+++ frontend/src/app/pages/admin/groups/groups.component.ts
@@ -10,6 +10,8 @@ import { MaterialModule } from '../../../material-module'
 import { ADMIN_GROUP } from '@shared/constants'
 import { RouterLink } from '@angular/router'
 import { SpinnerService } from '../../../services/spinner.service'
+import { MatDialog } from '@angular/material/dialog'
+import { ConfirmComponent } from '../../../dialogs/confirm/confirm.component'
 
 @Component({
   selector: 'app-groups',
@@ -41,6 +43,7 @@ export class GroupsComponent {
   private adminService = inject(AdminService)
   private snackbarService = inject(SnackbarService)
   private spinnerService = inject(SpinnerService)
+  private dialog = inject(MatDialog)
 
   async ngAfterViewInit() {
     try {
@@ -54,16 +57,31 @@ export class GroupsComponent {
     }
   }
 
-  async delete(id: string) {
-    try {
-      this.spinnerService.show()
-      await this.adminService.deleteGroup(id)
-      this.dataSource.data = this.dataSource.data.filter(g => g.id !== id)
-      this.snackbarService.show('Group was deleted.')
-    } catch (_e) {
-      this.snackbarService.error('Could not delete group.')
-    } finally {
-      this.spinnerService.hide()
-    }
+  delete(id: string) {
+    const group = this.dataSource.data.find(g => g.id === id)
+    const dialogRef = this.dialog.open(ConfirmComponent, {
+      data: {
+        message: `Are you sure you want to remove group '${group?.name ?? id}'?`,
+        header: 'Delete',
+      },
+    })
+
+    dialogRef.afterClosed().subscribe(async (result) => {
+      if (!result) {
+        this.snackbarService.error('Group delete cancelled.')
+        return
+      }
+
+      try {
+        this.spinnerService.show()
+        await this.adminService.deleteGroup(id)
+        this.dataSource.data = this.dataSource.data.filter(g => g.id !== id)
+        this.snackbarService.message('Group was deleted.')
+      } catch (_e) {
+        this.snackbarService.error('Could not delete group.')
+      } finally {
+        this.spinnerService.hide()
+      }
+    })
   }
 }
